requirements.txt
torch
pickle
numpy
Pillow
torchvision
matplotlib
nltk 